requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.2.0
braceexpand==0.1.7
captum==0.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
clip-benchmark==1.6.1
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.42
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.11.0
lime==0.2.0.1
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
open-clip-torch==2.24.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
pycocoevalcap==1.2
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.5
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==1.43.0
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
submitit==1.4.5
sympy==1.12
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.16
tokenizers==0.15.2
torch==2.2.1
torchmetrics==1.3.2
torchvision==0.17.1
tqdm==4.66.2
transformers==4.39.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.4
wcwidth==0.2.13
webdataset @ git+https://github.com/tmbdev/webdataset.git@f11fd66c163722c607ec99475a6f3cb880ec35b8
wheel==0.41.2
yarl==1.9.4
